Add files via upload
Coding project to create game variation of the original tic tac toe. 
This variation allows each player to only lay a maximum of 3 player symbol on the empty space among the 9 boxes. If a player symbol is about to exceed filling 3 boxes, the latest player input will replace the 1st input by making the 1st input box to become empty. When the first player manages to fill 3 boxes vertically, horizontal, or diagonal line, declare that player as the winner.